♻️ EventForm 컴포넌트 분리
- App.tsx에서 이벤트 폼 관련 로직을 새로운 EventForm 컴포넌트로 추출
- 현재 props 전달이 과한 상태이므로 추후 전역 상태 관리를 통해 상태값 제어 필요
