+++ src/components/EventForm.tsx
@@ -1,0 +1,251 @@
+import {
+  Button,
+  Checkbox,
+  FormControl,
+  FormLabel,
+  Heading,
+  HStack,
+  Input,
+  Select,
+  Tooltip,
+  useToast,
+  VStack,
+} from '@chakra-ui/react';
+import { ChangeEvent } from 'react';
+
+import { CATEGORIES, NOTIFICATION_OPTIONS } from '../constants';
+import type { Event, RepeatType, EventForm as TEventForm } from '../types';
+import { findOverlappingEvents } from '../utils/eventOverlap';
+import { getTimeErrorMessage } from '../utils/timeValidation';
+
+interface Props {
+  events: Event[];
+  eventForm: TEventForm & {
+    isRepeating: boolean;
+    startTimeError: string | null;
+    endTimeError: string | null;
+  };
+  editingEvent: Event | null;
+  eventHandlers: {
+    setTitle: (title: string) => void;
+    setDate: (date: string) => void;
+    setDescription: (description: string) => void;
+    setLocation: (location: string) => void;
+    setCategory: (category: string) => void;
+    setNotificationTime: (notificationTime: number) => void;
+    setIsRepeating: (isRepeating: boolean) => void;
+    setOverlappingEvents: (events: Event[]) => void;
+    setIsOverlapDialogOpen: (isOpen: boolean) => void;
+    setRepeatType: (repeatType: RepeatType) => void;
+    setRepeatInterval: (repeatInterval: number) => void;
+    setRepeatEndDate: (repeatEndDate: string) => void;
+    handleStartTimeChange: (e: ChangeEvent<HTMLInputElement>) => void;
+    handleEndTimeChange: (e: ChangeEvent<HTMLInputElement>) => void;
+    saveEvent: (event: Event) => Promise<void>;
+    resetForm: () => void;
+  };
+}
+export const EventForm = ({ events, eventForm, eventHandlers, editingEvent }: Props) => {
+  const toast = useToast();
+  const {
+    title,
+    date,
+    startTime,
+    endTime,
+    description,
+    location,
+    category,
+    notificationTime,
+    isRepeating,
+    startTimeError,
+    endTimeError,
+  } = eventForm;
+  const {
+    setTitle,
+    setDate,
+    setDescription,
+    setLocation,
+    setCategory,
+    setNotificationTime,
+    setIsRepeating,
+    setRepeatType,
+    setRepeatInterval,
+    setRepeatEndDate,
+    setOverlappingEvents,
+    setIsOverlapDialogOpen,
+    handleStartTimeChange,
+    handleEndTimeChange,
+    saveEvent,
+    resetForm,
+  } = eventHandlers;
+
+  const addOrUpdateEvent = async () => {
+    if (!title || !date || !startTime || !endTime) {
+      toast({
+        title: '필수 정보를 모두 입력해주세요.',
+        status: 'error',
+        duration: 3000,
+        isClosable: true,
+      });
+      return;
+    }
+
+    if (startTimeError || endTimeError) {
+      toast({
+        title: '시간 설정을 확인해주세요.',
+        status: 'error',
+        duration: 3000,
+        isClosable: true,
+      });
+      return;
+    }
+
+    const eventData: Event | TEventForm = {
+      id: editingEvent ? editingEvent.id : undefined,
+      title,
+      date,
+      startTime,
+      endTime,
+      description,
+      location,
+      category,
+      repeat: eventForm.repeat,
+      notificationTime,
+    };
+
+    const overlapping = findOverlappingEvents(eventData, events);
+    if (overlapping.length > 0) {
+      setOverlappingEvents(overlapping);
+      setIsOverlapDialogOpen(true);
+    } else {
+      await saveEvent(eventData as Event);
+      resetForm();
+    }
+  };
+
+  return (
+    <VStack w='400px' spacing={5} align='stretch'>
+      <Heading>{editingEvent ? '일정 수정' : '일정 추가'}</Heading>
+
+      <FormControl>
+        <FormLabel>제목</FormLabel>
+        <Input value={title} onChange={(e) => setTitle(e.target.value)} />
+      </FormControl>
+
+      <FormControl>
+        <FormLabel>날짜</FormLabel>
+        <Input type='date' value={date} onChange={(e) => setDate(e.target.value)} />
+      </FormControl>
+
+      <HStack width='100%'>
+        <FormControl>
+          <FormLabel>시작 시간</FormLabel>
+          <Tooltip label={startTimeError} isOpen={!!startTimeError} placement='top'>
+            <Input
+              type='time'
+              value={startTime}
+              onChange={handleStartTimeChange}
+              onBlur={() => getTimeErrorMessage(startTime, endTime)}
+              isInvalid={!!startTimeError}
+            />
+          </Tooltip>
+        </FormControl>
+        <FormControl>
+          <FormLabel>종료 시간</FormLabel>
+          <Tooltip label={endTimeError} isOpen={!!endTimeError} placement='top'>
+            <Input
+              type='time'
+              value={endTime}
+              onChange={handleEndTimeChange}
+              onBlur={() => getTimeErrorMessage(startTime, endTime)}
+              isInvalid={!!endTimeError}
+            />
+          </Tooltip>
+        </FormControl>
+      </HStack>
+
+      <FormControl>
+        <FormLabel>설명</FormLabel>
+        <Input value={description} onChange={(e) => setDescription(e.target.value)} />
+      </FormControl>
+
+      <FormControl>
+        <FormLabel>위치</FormLabel>
+        <Input value={location} onChange={(e) => setLocation(e.target.value)} />
+      </FormControl>
+
+      <FormControl>
+        <FormLabel>카테고리</FormLabel>
+        <Select value={category} onChange={(e) => setCategory(e.target.value)}>
+          <option value=''>카테고리 선택</option>
+          {CATEGORIES.map((cat) => (
+            <option key={cat} value={cat}>
+              {cat}
+            </option>
+          ))}
+        </Select>
+      </FormControl>
+
+      <FormControl>
+        <FormLabel>반복 설정</FormLabel>
+        <Checkbox isChecked={isRepeating} onChange={(e) => setIsRepeating(e.target.checked)}>
+          반복 일정
+        </Checkbox>
+      </FormControl>
+
+      <FormControl>
+        <FormLabel>알림 설정</FormLabel>
+        <Select
+          value={notificationTime}
+          onChange={(e) => setNotificationTime(Number(e.target.value))}
+        >
+          {NOTIFICATION_OPTIONS.map((option) => (
+            <option key={option.value} value={option.value}>
+              {option.label}
+            </option>
+          ))}
+        </Select>
+      </FormControl>
+
+      {isRepeating && (
+        <VStack width='100%'>
+          <FormControl>
+            <FormLabel>반복 유형</FormLabel>
+            <Select
+              value={eventForm.repeat.type}
+              onChange={(e) => setRepeatType(e.target.value as RepeatType)}
+            >
+              <option value='daily'>매일</option>
+              <option value='weekly'>매주</option>
+              <option value='monthly'>매월</option>
+              <option value='yearly'>매년</option>
+            </Select>
+          </FormControl>
+          <HStack width='100%'>
+            <FormControl>
+              <FormLabel>반복 간격</FormLabel>
+              <Input
+                type='number'
+                value={eventForm.repeat.interval}
+                onChange={(e) => setRepeatInterval(Number(e.target.value))}
+                min={1}
+              />
+            </FormControl>
+            <FormControl>
+              <FormLabel>반복 종료일</FormLabel>
+              <Input
+                type='date'
+                value={eventForm.repeat.endDate}
+                onChange={(e) => setRepeatEndDate(e.target.value)}
+              />
+            </FormControl>
+          </HStack>
+        </VStack>
+      )}
+
+      <Button data-testid='event-submit-button' onClick={addOrUpdateEvent} colorScheme='blue'>
+        {editingEvent ? '일정 수정' : '일정 추가'}
+      </Button>
+    </VStack>
+  );
+};
